Update HKCU"\...\data value to accomodate Office 365 ProPlus; add Break in Foreach
It appears that in the ProPlus version of Office *365*, the binary values for the "data" reg key are slightly different. 4 is still Dark, but 5 is Light (and 7 is Colorful, if you're curious). I added an If block to check if Office365ProPlus is installed, and if so, changed that value. I also added a Break in the Foreach loop once the valid key is found.